Move image fetching to own file
